Second pass at setting order of stories
Still not actually as good as deciding on a container for views or
providing menu authoring, but a little less of a hack.

Add optional temporary vm:asOrdinal property for stories which defines
the order they should be presented in. Sort the output with that as the
first key.

Add n variable to transforms which contains the 1-based row number.

Replace old hacks for economic system story ordering.