Reshuffle RIS3 views as stories
Fits better with the vision for this map, though makes the transform
less pleasant to work with. Numbering scheme copied from the sheet.
Notice that some triples apply only to stories, and others only to
storypoints.

Split tile_id and geog_id inputs again, as all storypoints that show
labels must have that named linked, but may not have tiles.