Add _cross_cols for applying logic per row and col
Rough attempt at enabling transform on a sheet that is not a tidy
observation per row. So, triples will be processed per-cell when
specific field of columns is supplied.

For now list of columns in sheet must still be exhaustive, but avoid
repeating triples lines for each column in this case. Also add
conditional that allows for logic on the cell contents but can give
column name as the result value.