Refine behaviour of literals with lang tags
When loading from a json model, there is logic to detect languages
encoded onto object literals. This does not remove the json level
encoding but duplicates it, so on serialisation as ttl for example
results in multiple tags. Instead strip the value on load, and on
serialisation to json reapply the json form.

There is some slightly magical logic outside of rdf conventions here
where encoding a language for javascript array and object values,
required for labels in multiple languages.