Relax minimum Python version and gate both ends
Can match the minimum Python version support to that of rdflib rather
than requiring the latest.

Update various dependencies and the use of poetry in the Makefile.

Gate on both Python 3.8 and 3.11 to ensure compatibility.

Also remove an unused requirements.txt and a random pyc that was
committed by mistake.